+++ cw-vault-standard/src/extensions/force_unlock.rs
@@ -1,5 +1,5 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{to_binary, Coin, CosmosMsg, StdResult, Uint128, WasmMsg};
+use cosmwasm_std::{to_json_binary, Coin, CosmosMsg, StdResult, Uint128, WasmMsg};
 
 use crate::{ExtensionExecuteMsg, VaultStandardExecuteMsg};
 
@@ -15,8 +15,6 @@ pub enum ForceUnlockExecuteMsg {
         /// The address which should receive the withdrawn assets. If not set,
         /// the caller address will be used instead.
         recipient: Option<String>,
-        /// The amount of vault tokens to force redeem.
-        amount: Uint128,
     },
 
     /// Force withdraw from a position that is already unlocking (Unlock has
@@ -47,7 +45,7 @@ impl ForceUnlockExecuteMsg {
     pub fn into_cosmos_msg(self, contract_addr: String, funds: Vec<Coin>) -> StdResult<CosmosMsg> {
         Ok(WasmMsg::Execute {
             contract_addr,
-            msg: to_binary(&VaultStandardExecuteMsg::VaultExtension(
+            msg: to_json_binary(&VaultStandardExecuteMsg::VaultExtension(
                 ExtensionExecuteMsg::ForceUnlock(self),
             ))?,
             funds,