+++ cw-vault-standard/src/extensions/lockup.rs
@@ -1,5 +1,5 @@
 use cosmwasm_schema::{cw_serde, QueryResponses};
-use cosmwasm_std::{to_binary, Addr, Coin, CosmosMsg, StdResult, Uint128, WasmMsg};
+use cosmwasm_std::{to_json_binary, Addr, Coin, CosmosMsg, StdResult, Uint128, WasmMsg};
 use cw_utils::{Duration, Expiration};
 
 use crate::{ExtensionExecuteMsg, VaultStandardExecuteMsg};
@@ -22,10 +22,7 @@ pub enum LockupExecuteMsg {
     ///
     /// Like Redeem, this takes an amount so that the same API can be used for
     /// CW4626 and native tokens.
-    Unlock {
-        /// The amount of vault tokens to unlock.
-        amount: Uint128,
-    },
+    Unlock {},
 
     /// EmergencyUnlock is called to initiate unlocking a locked position held
     /// by the vault.
@@ -54,7 +51,7 @@ impl LockupExecuteMsg {
     pub fn into_cosmos_msg(self, contract_addr: String, funds: Vec<Coin>) -> StdResult<CosmosMsg> {
         Ok(WasmMsg::Execute {
             contract_addr,
-            msg: to_binary(&VaultStandardExecuteMsg::VaultExtension(
+            msg: to_json_binary(&VaultStandardExecuteMsg::VaultExtension(
                 ExtensionExecuteMsg::Lockup(self),
             ))?,
             funds,