+++ cw-vault-standard/src/helper.rs
@@ -2,7 +2,8 @@ use std::marker::PhantomData;
 
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{
-    coin, to_binary, Addr, Api, CosmosMsg, QuerierWrapper, StdResult, Uint128, WasmMsg,
+    coin, to_json_binary, Addr, CosmosMsg, Decimal, Deps, QuerierWrapper, StdResult, Uint128,
+    WasmMsg,
 };
 use schemars::JsonSchema;
 use serde::Serialize;
@@ -39,8 +40,8 @@ where
 
     /// Check the address against the api and return a checked version of the
     /// struct.
-    pub fn check(&self, api: &dyn Api) -> StdResult<VaultContract<E, Q>> {
-        Ok(VaultContract::new(&api.addr_validate(&self.addr)?))
+    pub fn check(&self, deps: Deps) -> StdResult<VaultContract<E, Q>> {
+        VaultContract::new(&deps.querier, &deps.api.addr_validate(&self.addr)?)
     }
 }
 
@@ -50,6 +51,10 @@ where
 pub struct VaultContract<E = ExtensionExecuteMsg, Q = ExtensionQueryMsg> {
     /// The address of the vault contract.
     pub addr: Addr,
+    /// The base token of the vault contract.
+    pub base_token: String,
+    /// The vault token denom of the vault contract.
+    pub vault_token: String,
     /// The extension enum for ExecuteMsg variants.
     execute_msg_extension: PhantomData<E>,
     /// The extension enum for QueryMsg variants.
@@ -62,39 +67,32 @@ where
     Q: Serialize + JsonSchema,
 {
     /// Create a new VaultContract instance.
-    pub fn new(addr: &Addr) -> Self {
-        Self {
+    pub fn new(querier: &QuerierWrapper, addr: &Addr) -> StdResult<Self> {
+        // Query vault info
+        let vault_info: VaultInfoResponse =
+            querier.query_wasm_smart(addr, &VaultStandardQueryMsg::<Q>::Info {})?;
+
+        Ok(Self {
             addr: addr.clone(),
+            base_token: vault_info.base_token,
+            vault_token: vault_info.vault_token,
             execute_msg_extension: PhantomData,
             query_msg_extension: PhantomData,
-        }
+        })
     }
 
     /// Returns a CosmosMsg to deposit base tokens into the vault.
     pub fn deposit(
         &self,
         amount: impl Into<Uint128>,
-        base_denom: &str,
         recipient: Option<String>,
     ) -> StdResult<CosmosMsg> {
         let amount = amount.into();
 
         Ok(WasmMsg::Execute {
             contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Deposit { amount, recipient })?,
-            funds: vec![coin(amount.u128(), base_denom)],
-        }
-        .into())
-    }
-
-    /// Returns a CosmosMsg to deposit tokens into the vault, leaving the native
-    /// funds field empty. This is useful for depositing cw20 tokens. The
-    /// caller should have approved spend for the cw20 tokens first.
-    pub fn deposit_cw20(&self, amount: Uint128, recipient: Option<String>) -> StdResult<CosmosMsg> {
-        Ok(WasmMsg::Execute {
-            contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Deposit { amount, recipient })?,
-            funds: vec![],
+            msg: to_json_binary(&VaultStandardExecuteMsg::<E>::Deposit { recipient })?,
+            funds: vec![coin(amount.u128(), &self.base_token)],
         }
         .into())
     }
@@ -103,14 +101,13 @@ where
     pub fn redeem(
         &self,
         amount: impl Into<Uint128>,
-        vault_token_denom: &str,
         recipient: Option<String>,
     ) -> StdResult<CosmosMsg> {
         let amount = amount.into();
         Ok(WasmMsg::Execute {
             contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Redeem { amount, recipient })?,
-            funds: vec![coin(amount.u128(), vault_token_denom)],
+            msg: to_json_binary(&VaultStandardExecuteMsg::<E>::Redeem { recipient })?,
+            funds: vec![coin(amount.u128(), &self.vault_token)],
         }
         .into())
     }
@@ -131,34 +128,6 @@ where
         querier.query_wasm_smart(&self.addr, &VaultStandardQueryMsg::<Q>::Info {})
     }
 
-    /// Queries the vault for a preview of a deposit
-    pub fn query_preview_deposit(
-        &self,
-        querier: &QuerierWrapper,
-        amount: impl Into<Uint128>,
-    ) -> StdResult<Uint128> {
-        querier.query_wasm_smart(
-            &self.addr,
-            &VaultStandardQueryMsg::<Q>::PreviewDeposit {
-                amount: amount.into(),
-            },
-        )
-    }
-
-    /// Queries the vault for a preview of a redeem
-    pub fn query_preview_redeem(
-        &self,
-        querier: &QuerierWrapper,
-        amount: impl Into<Uint128>,
-    ) -> StdResult<Uint128> {
-        querier.query_wasm_smart(
-            &self.addr,
-            &VaultStandardQueryMsg::<Q>::PreviewRedeem {
-                amount: amount.into(),
-            },
-        )
-    }
-
     /// Queries the vault for the total assets held in the vault
     pub fn query_total_assets(&self, querier: &QuerierWrapper) -> StdResult<Uint128> {
         querier.query_wasm_smart(&self.addr, &VaultStandardQueryMsg::<Q>::TotalAssets {})
@@ -169,6 +138,18 @@ where
         querier.query_wasm_smart(
             &self.addr,
             &VaultStandardQueryMsg::<Q>::TotalVaultTokenSupply {},
+        )
+    }
+
+    /// Queries the vault for the vault token exchange rate
+    pub fn query_vault_token_exchange_rate(
+        &self,
+        quote_denom: String,
+        querier: &QuerierWrapper,
+    ) -> StdResult<Decimal> {
+        querier.query_wasm_smart(
+            &self.addr,
+            &VaultStandardQueryMsg::<Q>::VaultTokenExchangeRate { quote_denom },
         )
     }